pyproject.toml
[project]
name = "LexTrade"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn>=0.27.0",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "sqlalchemy-utils (>=0.41.2,<0.42.0)",
    "sqlacodegen (==3.0.0rc5)",
    "aiomysql>=0.2.0",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "aiosmtplib>=3.0.1",
    "email-validator>=2.1.0",
    "alibabacloud-dysmsapi20170525 (>=3.1.1,<4.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "jwt (>=1.3.1,<2.0.0)",
    "phonenumbers (>=9.0.1,<10.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
